1. 거래일지 목록 호출 API 작업 commit 2. 거래일지 DB 설계 작업 commit

+++ src/main/java/com/jbproject/jutopia/rest/entity/TradeCorpDetailEntity.java
@@ -26,12 +26,12 @@ public class TradeCorpDetailEntity extends BaseEntity{
     private LocalDateTime investDate;
 
     private Long buyQuantity;
-    private Long buyAmount;
+    private Double buyAmount;
     private Double buyFee;
     private Long sellQuantity;
-    private Long sellAmount;
+    private Double sellAmount;
     private Double sellFee;
-    private Long profitLossAmount;
+    private Double profitLossAmount;
 
 
     @ManyToOne(fetch = FetchType.LAZY)
